+++ srcs/HttpResponse/DELETE/Readme.Md
@@ -11,3 +11,12 @@
 	+ それらの行動をする際に対象のファイルがどういう状態かを確認する  
 	+ 状態に沿って正しいステータスコード を設定する  
 みたいな感じの理解を持っている  
+  
+１　ヘッダーを確認  
+　　　　必要なヘッダーがなければ400  
+２　対象のファイル確認  
+　　　　ファイルが存在しなければ404  
+３　削除可能かどうか（権限はどうか  
+　　　　権限が削除可能でなければ403  
+４　削除　200もしくは204を返す  
+
